pyproject.toml
[tool.poetry]
name = "rasapro-3-9-x-assistant"
version = "0.1.0"
description = "A simple Rasa assistant"
authors = ["John Trammell <j.trammell@rasa.com>"]
license = "Apache-2.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.8.1,<3.11"
rasa-pro = { version = "3.9.5", source = "rasa-pro-python" }
psycopg = { version = "^3.2.1", extras = ["binary", "pool"] }

[[tool.poetry.source]]
name = "rasa-pro-python"
url = "https://europe-west3-python.pkg.dev/rasa-releases/rasa-pro-python/simple"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
